initial commit, practice programs from class.
initial commit, these are the practice programs from class. They are in whatever state they were in at the end of class. These are the versions I wrote on my local computer, and not the versions that were submitted to the school's grading software.  These are practice programs, and tests.